chore: example app expo router + native ui
