+++ example/app/(tabs)/(home)/index.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import ToastDemo from '../../../components/toast-demo';
+
+const HomeScreen: React.FC = () => {
+  return <ToastDemo />;
+};
+
+export default HomeScreen;
