+++ example/app/(tabs)/_layout.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import { Icon, Label, NativeTabs } from 'expo-router/unstable-native-tabs';
+
+const TabsLayout: React.FC = () => {
+  return (
+    <NativeTabs>
+      <NativeTabs.Trigger name="(home)">
+        <Label>Home</Label>
+        <Icon sf="house.fill" drawable="custom_android_drawable" />
+      </NativeTabs.Trigger>
+      <NativeTabs.Trigger name="(settings)">
+        <Icon sf="gear" drawable="custom_settings_drawable" />
+        <Label>Settings</Label>
+      </NativeTabs.Trigger>
+    </NativeTabs>
+  );
+};
+
+export default TabsLayout;
